marvin.py
The script below gives the well-known Ev3rstorm robot a new and rather obnoxious personality, 
that of Marvin, the paranoid android star of the Hitchhiker's Guide to the Galaxy. 
This script was kindly offered by Denis Demidov, contributor to the EV3dev project and Senior Researcher 
at the Supercomputer Center of Russian Academy of Sciences. It is intended to animate the well-known 
Ev3rstorm robot, build instructions for which can be found HERE. These instructions are designed 
for the home edition of the EV3 kit - if you have the education edition then you will probably not be 
able to use this script because it makes use of the remote control and the ball-shooter. 
The essential characteristics of this robot are that it is driven by large motors attached to ports B and C, 
that a medium motor is attached to port A or D, that the touch sensor, colour sensor and IR sensors 
are attached to any sensor ports. This script is copied from this page.

Interestingly, the script multiplies the number of functions that can be activated by the remote control 
by using two different channels. With the remote control set to channel one, the remote control read and 
blue button pairs control the large motors, making the robot move around. 
With the remote control set to channel 2, either of the remote control 'up' buttons will fire a ball up 
and either of the remote control 'down' buttons will fire a ball down.